Revert "Change the number from float to double"
This reverts commit 76dc096042bf9bc91ca4143e41bb3d14f35ed06e.

+++ components/taixia/number/taixia_number.cpp
@@ -18,7 +18,7 @@ static const char *const TAG = "taixia.nubmer";
       this->parent_->send(6, 0, 0, SERVICE_ID_READ_STATUS, 0xffff);
   }
 
-  void TaiXiaNumber::control(double value) {
+  void TaiXiaNumber::control(float value) {
     if (this->service_id_ > 0) {
       this->parent_->set_number(this->sa_id_, this->service_id_, value);
       this->publish_state(value);
